Fixed a few bugs. Add more bug benchmark scripts. Refactor python scripts to better handle knobs.

+++ script/maple/benchmark/aget_bug2_training.py
@@ -16,23 +16,34 @@ Authors - Jie Yu (jieyu@umich.edu)
 """
 
 import os
+import subprocess
 from maple.core import config
+from maple.core import logging
 from maple.core import testing
 
 _sio = [None, os.devnull, 'stderr']
 
-class Test(testing.CmdlineTest):
+class Test(testing.ServerTest):
     def __init__(self, input_idx):
-        testing.CmdlineTest.__init__(self, input_idx)
-        #self.add_input(([self.bin(), '-n1', 'http://apache.cyberuse.com/httpd/httpd-2.2.21.tar.gz', '-l', 'aget.file'], _sio))
-        self.add_input(([self.bin(), '-n1', 'http://zodiac.eecs.umich.edu/httpd-2.2.21.tar.gz', '-l', 'aget.file'], _sio))
-    def bin(self):
-        return config.benchmark_home('aget_bug1') + '/aget'
+        testing.ServerTest.__init__(self, input_idx)
+        self.add_input(([self.bin(), '-n1', 'http://apache.cyberuse.com/httpd/httpd-2.2.21.tar.gz', '-l', 'aget.file'], _sio))
     def setup(self):
-        assert not os.path.exists('aget.file')
+        if os.path.exists('aget.file'):
+            os.remove('aget.file')
     def tear_down(self):
-        assert os.path.exists('aget.file')
-        os.remove('aget.file')
+        if os.path.exists('aget.file'):
+            os.remove('aget.file')
+    def start(self):
+        args, sio = self.input()
+        cmd = []
+        if self.prefix != None:
+            cmd.extend(self.prefix)
+        cmd.extend(args)
+        self.proc = subprocess.Popen(cmd)
+    def stop(self):
+        self.proc.wait()
+    def bin(self):
+        return config.benchmark_home('aget_bug2') + '/aget'
 
 def get_test(input_idx='default'):
     return Test(input_idx)
